refactor: rename common.js to ignore-pipe-errors.js
