+++ src/dredd-command.js
@@ -7,10 +7,10 @@ const spawnSync = require('cross-spawn').sync;
 
 const configUtils = require('./config-utils');
 const Dredd = require('./dredd');
+const ignorePipeErrors = require('./ignore-pipe-errors');
 const interactiveConfig = require('./interactive-config');
 const logger = require('./logger');
 const { applyLoggingOptions } = require('./configuration');
-const { ignorePipeErrors } = require('./common');
 const { spawn } = require('./child-process');
 
 const packageData = require('../package.json');
