+++ src/ignore-pipe-errors.js
@@ -1,7 +1,5 @@
-function ignorePipeErrors(proc) {
+module.exports = function ignorePipeErrors(proc) {
   proc.stdin.on('error', () => {});
   proc.stdout.on('error', () => {});
   proc.stderr.on('error', () => {});
-}
-
-module.exports = { ignorePipeErrors };
+};
